+++ sc/social/like/testing.py
@@ -46,9 +46,7 @@ class Fixture(PloneSandboxLayer):
             self.applyProfile(portal, 'collective.cover:default')
 
         self.applyProfile(portal, 'sc.social.like:default')
-        portal_workflow = portal.portal_workflow
-        portal_workflow.setChainForPortalTypes(
-            ['Document'], ['simple_publication_workflow'])
+        portal.portal_workflow.setDefaultChain('simple_publication_workflow')
 
 FIXTURE = Fixture()